Moves MapConfig to models directory
