+++ lib/windshaft/models/mapconfig.js
