+++ lib/windshaft/server.js
@@ -3,7 +3,7 @@ var   express     = require('express')
     , RedisPool   = require('redis-mpool')
     , LocalizedResourcePurger = require('./cache/localized_resource_purger')
     , MapStore    = require('./storages/mapstore')
-    , MapConfig   = require('./mapconfig')
+    , MapConfig   = require('./models/mapconfig')
     , _           = require('underscore')
     , mapnik      = require('mapnik')
     , Step        = require('step')
