+++ lib/windshaft/storages/mapstore.js
@@ -3,7 +3,7 @@
 var Step       = require('step');
 var RedisPool  = require('redis-mpool');
 var _          = require('underscore');
-var MapConfig  = require('../mapconfig');
+var MapConfig  = require('../models/mapconfig');
 
 /// API: Create MapStore 
 //
